+++ src/utilities/hypre_memory.h
@@ -1,28 +1,15 @@
 /*BHEADER**********************************************************************
- * Copyright (c) 2006   The Regents of the University of California.
+ * Copyright (c) 2008,  Lawrence Livermore National Security, LLC.
  * Produced at the Lawrence Livermore National Laboratory.
- * Written by the HYPRE team, UCRL-CODE-222953.
- * All rights reserved.
+ * This file is part of HYPRE.  See file COPYRIGHT for details.
  *
- * This file is part of HYPRE (see http://www.llnl.gov/CASC/hypre/).
- * Please see the COPYRIGHT_and_LICENSE file for the copyright notice, 
- * disclaimer and the GNU Lesser General Public License.
- *
- * This program is free software; you can redistribute it and/or modify it
- * under the terms of the GNU General Public License (as published by the Free
+ * HYPRE is free software; you can redistribute it and/or modify it under the
+ * terms of the GNU Lesser General Public License (as published by the Free
  * Software Foundation) version 2.1 dated February 1999.
  *
- * This program is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the IMPLIED WARRANTY OF MERCHANTABILITY or
- * FITNESS FOR A PARTICULAR PURPOSE.  See the terms and conditions of the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU Lesser General Public License
- * along with this program; if not, write to the Free Software Foundation,
- * Inc., 59 Temple Place, Suite 330, Boston, MA 02111-1307 USA
- *
- * $Revision: 2.3 $
+ * $Revision: 2.6 $
  ***********************************************************************EHEADER*/
+
 
 /******************************************************************************
  *
@@ -32,6 +19,9 @@
 
 #ifndef hypre_MEMORY_HEADER
 #define hypre_MEMORY_HEADER
+
+#include <stdio.h>
+#include <stdlib.h>
 
 #ifdef __cplusplus
 extern "C" {
@@ -72,13 +62,13 @@ extern "C" {
 #define hypre_FinalizeMemoryDebug()  
 
 #define hypre_TAlloc(type, count) \
-( (type *)hypre_MAlloc((unsigned int)(sizeof(type) * (count))) )
+( (type *)hypre_MAlloc((size_t)(sizeof(type) * (count))) )
 
 #define hypre_CTAlloc(type, count) \
-( (type *)hypre_CAlloc((unsigned int)(count), (unsigned int)sizeof(type)) )
+( (type *)hypre_CAlloc((size_t)(count), (size_t)sizeof(type)) )
 
 #define hypre_TReAlloc(ptr, type, count) \
-( (type *)hypre_ReAlloc((char *)ptr, (unsigned int)(sizeof(type) * (count))) )
+( (type *)hypre_ReAlloc((char *)ptr, (size_t)(sizeof(type) * (count))) )
 
 #define hypre_TFree(ptr) \
 ( hypre_Free((char *)ptr), ptr = NULL )
@@ -89,16 +79,16 @@ extern "C" {
 #ifdef HYPRE_USE_PTHREADS
 
 #define hypre_SharedTAlloc(type, count) \
-( (type *)hypre_SharedMAlloc((unsigned int)(sizeof(type) * (count))) )
+( (type *)hypre_SharedMAlloc((size_t)(sizeof(type) * (count))) )
 
 
 #define hypre_SharedCTAlloc(type, count) \
-( (type *)hypre_SharedCAlloc((unsigned int)(count),\
-                             (unsigned int)sizeof(type)) )
+( (type *)hypre_SharedCAlloc((size_t)(count),\
+                             (size_t)sizeof(type)) )
 
 #define hypre_SharedTReAlloc(ptr, type, count) \
 ( (type *)hypre_SharedReAlloc((char *)ptr,\
-                              (unsigned int)(sizeof(type) * (count))) )
+                              (size_t)(sizeof(type) * (count))) )
 
 #define hypre_SharedTFree(ptr) \
 ( hypre_SharedFree((char *)ptr), ptr = NULL )
@@ -117,16 +107,16 @@ extern "C" {
  *--------------------------------------------------------------------------*/
 
 /* hypre_memory.c */
-int hypre_OutOfMemory( int size );
-char *hypre_MAlloc( int size );
-char *hypre_CAlloc( int count , int elt_size );
-char *hypre_ReAlloc( char *ptr , int size );
-void hypre_Free( char *ptr );
-char *hypre_SharedMAlloc( int size );
-char *hypre_SharedCAlloc( int count , int elt_size );
-char *hypre_SharedReAlloc( char *ptr , int size );
-void hypre_SharedFree( char *ptr );
-double *hypre_IncrementSharedDataPtr( double *ptr , int size );
+int hypre_OutOfMemory ( size_t size );
+char *hypre_MAlloc ( size_t size );
+char *hypre_CAlloc ( size_t count , size_t elt_size );
+char *hypre_ReAlloc ( char *ptr , size_t size );
+void hypre_Free ( char *ptr );
+char *hypre_SharedMAlloc ( size_t size );
+char *hypre_SharedCAlloc ( size_t count , size_t elt_size );
+char *hypre_SharedReAlloc ( char *ptr , size_t size );
+void hypre_SharedFree ( char *ptr );
+double *hypre_IncrementSharedDataPtr ( double *ptr , size_t size );
 
 /* memory_dmalloc.c */
 int hypre_InitMemoryDebugDML( int id );
